feat(#1532): CLEAN: split dependency-existence-check.py (1488 lines)
